refactor: extract workspace scanning and opening to lib/workspaceIO.ts
Move scanDir, scanForHttpFiles, tryLoadEnvFiles, openFolderByPath and
safeJoinPath out of App.svelte into src/lib/workspaceIO.ts. The
UI-owned per-environment Key Vault cache stays in App.svelte and is
reached through KeyVaultHooks callbacks passed to openFolderByPath.
No behavior change.
